Add files via upload
This work presents the implementation and analysis
of the Ant Colony algorithm (ACO) to solve the
traveling salesman problem (TSP). The objective was to find the
shortest route between cities on a map, using the original calculation
ginal and Euclidean distance. The algorithm incorporates strategies
as short and long term memory, aspiration criterion
and pheromone evaporation. The experiments demonstrated
the effectiveness of the ACO in optimizing routes, obtaining results
satisfactory in terms of precision and performance, even for
large-scale instances, standing out as an approach
robust for NP-hard problems.